factor out JSON file reader.
	renamed:    lib/package_json_loader.js -> lib/json_reader.js
	modified:   lib/package_json_loader.js
